+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## [2.0.0]
+
+* Breaking change: ApiLogger renamed to ApiUtilsConfig
+* Added the ability to set a global timeout for all requests on ApiUtilsConfig
+* Added the ability to set a timeout per request. This will override the global setting if set.
+
 ## [1.0.0]
 
 * Documentation updates, stable release